Refactor the code

+++ src/main/java/com/uoc/ease_care_backend/service/AmbulanceService.java
@@ -2,15 +2,15 @@ package com.uoc.ease_care_backend.service;
 
 import com.google.api.core.ApiFuture;
 import com.google.cloud.firestore.*;
+import com.google.firebase.auth.DeleteUsersResult;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseAuthException;
+import com.google.firebase.auth.UserRecord;
 import com.google.firebase.cloud.FirestoreClient;
 import com.uoc.ease_care_backend.dto.AmbulanceDTO;
-import com.uoc.ease_care_backend.dto.ServiceProviderDTO;
-import org.apache.commons.codec.binary.StringUtils;
 import org.springframework.stereotype.Service;
 
-import java.lang.reflect.ParameterizedType;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
 import java.util.logging.Logger;
@@ -21,12 +21,22 @@ public class AmbulanceService {
     private Logger logger = Logger.getLogger(String.valueOf(ProviderService.class));
 
     public boolean registerAmbulance(AmbulanceDTO dto) {
-        Firestore dbFirestore = FirestoreClient.getFirestore();
-        String ambulanceID = String.valueOf(generateID());
-        dto.setAmbulanceId(ambulanceID);
-        ApiFuture<WriteResult> collectionsApiFuture = dbFirestore.collection(collectionName).document(ambulanceID).set(dto);
-        if(getAmbulance(ambulanceID) != null) {
-            return true;
+        UserRecord.CreateRequest req = new UserRecord.CreateRequest();
+        req.setEmail(dto.getEmail());
+        req.setPassword(dto.getPassword());
+        dto.setLastSignIn("-");
+        try {
+            UserRecord user = FirebaseAuth.getInstance().createUser(req);
+            Firestore dbFirestore = FirestoreClient.getFirestore();
+            String ambulanceID = String.valueOf(user.getUid());
+            dto.setUserId(ambulanceID);
+            ApiFuture<WriteResult> collectionsApiFuture = dbFirestore.collection(collectionName).document(ambulanceID).set(dto);
+            if(getAmbulance(ambulanceID) != null) {
+                return true;
+            }
+            return false;
+        } catch (FirebaseAuthException e) {
+            e.printStackTrace();
         }
         return false;
     }
@@ -43,15 +53,21 @@ public class AmbulanceService {
         }
         for (QueryDocumentSnapshot document : documents) {
             AmbulanceDTO dto=new AmbulanceDTO();
-            dto.setAmbulanceId(document.get("ambulanceId").toString());
+            dto.setUserId(document.get("userId").toString());
             dto.setServiceProviderId(document.get("serviceProviderId").toString());
-            dto.setVehicleNumber(document.get("vehicleNumber").toString());
-            dto.setDriverName(document.get("driverName").toString());
+            dto.setNumber(document.get("number").toString());
             dto.setDriverNIC(document.get("driverNIC").toString());
+            dto.setName(document.get("name").toString());
             dto.setContactNumber(document.get("contactNumber").toString());
-            dto.setUserName(document.get("userName").toString());
+            dto.setLastSignIn(document.get("lastSignIn").toString());
+            dto.setAmbulanceCharge(document.get("ambulanceCharge").toString());
+            dto.setEmail(document.get("email").toString());
             ambulances.add(dto);
         }
+
+
+
+
         return ambulances;
     }
 
@@ -75,14 +91,15 @@ public class AmbulanceService {
             document = future.get();
             if(document.exists()) {
                 AmbulanceDTO dto=new AmbulanceDTO();
-                dto.setAmbulanceId(document.get("ambulanceId").toString());
+                dto.setUserId(document.get("userId").toString());
                 dto.setServiceProviderId(document.get("serviceProviderId").toString());
-                dto.setVehicleNumber(document.get("vehicleNumber").toString());
-                dto.setDriverName(document.get("driverName").toString());
+                dto.setNumber(document.get("number").toString());
                 dto.setDriverNIC(document.get("driverNIC").toString());
+                dto.setName(document.get("name").toString());
                 dto.setContactNumber(document.get("contactNumber").toString());
-                dto.setUserName(document.get("userName").toString());
-                dto.setPassword(document.get("password").toString());
+                dto.setLastSignIn(document.get("lastSignIn").toString());
+                dto.setAmbulanceCharge(document.get("ambulanceCharge").toString());
+                dto.setEmail(document.get("email").toString());
                 return dto;
             }
         } catch (InterruptedException  | ExecutionException e) {
@@ -95,12 +112,14 @@ public class AmbulanceService {
 
     public boolean editAmbulance(AmbulanceDTO dto){
         Firestore dbFirestore = FirestoreClient.getFirestore();
-        ApiFuture<WriteResult> collectionsApiFuture = dbFirestore.collection(collectionName).document(dto.getAmbulanceId()).set(dto);
+        ApiFuture<WriteResult> collectionsApiFuture = dbFirestore.collection(collectionName).document(dto.getUserId()).set(dto);
         try {
-            return (collectionsApiFuture.get().getUpdateTime().toString() == null) ? false :true;
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        } catch (ExecutionException e) {
+            UserRecord.UpdateRequest req = new UserRecord.UpdateRequest(dto.getUserId());
+            req.setEmail(dto.getEmail());
+            req.setPassword(dto.getPassword());
+            UserRecord updateUser = FirebaseAuth.getInstance().updateUser(req);
+            return true;
+        } catch (FirebaseAuthException e) {
             e.printStackTrace();
         }
         return false;
@@ -109,24 +128,20 @@ public class AmbulanceService {
     public boolean deleteAmbulance(String ambulanceId){
         Firestore dbFirestore = FirestoreClient.getFirestore();
         if(getAmbulance(ambulanceId) != null) {
-            ApiFuture<WriteResult> writeResult = dbFirestore.collection(collectionName).document(ambulanceId).delete();
-            return true;
+            try {
+                List<String> uids = new ArrayList<>();
+                uids.add(ambulanceId);
+                DeleteUsersResult deleteUsersResult = FirebaseAuth.getInstance().deleteUsers(uids);
+                if (deleteUsersResult.getSuccessCount()>0){
+                    ApiFuture<WriteResult> writeResult = dbFirestore.collection(collectionName).document(ambulanceId).delete();
+                    return true;
+                }
+            } catch (FirebaseAuthException e) {
+                e.printStackTrace();
+            }
+            return false;
         }else {
             return false;
         }
-
-
-    }
-
-    public int generateID(){
-        List<AmbulanceDTO> list =getAmbulances();
-        int lastID;
-        if(list.size()>0){
-            lastID = Integer.parseInt(list.get(list.size()-1).getAmbulanceId());
-        }else{
-            lastID =0;
-        }
-
-        return lastID+1;
     }
 }
